requirements.txt
requests
psycopg2-binary
flask==2.2.5